src/monitors/requirements.txt
PyYAML==3.10
pyzmq==13.1.0
redis==2.10.3
requests
cffi>=0.8
cryptography
pyOpenSSL
ndg-httpsclient
pyasn1
fabric
